// src/lib/api/schema.ts
export interface components {
    schemas: {
// ...
        Toggle: {
            /** @example __TEST__ */
            name: string;
            /** @example false */
            toggle: string;
            /** @example some text */
            notification: string;
        };
// ...
    };
    responses: {
        /** @description Access token is missing or invalid; try to refresh your token */
        UnauthorizedError: {
            headers: {
                [name: string]: unknown;
            };
            content?: never;
        };
    };
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}